move NetSuite mocks into one folder
