+++ test/ss2/ns/error.js
