+++ test/ss2/ns/log.js
