+++ test/ss2/ns/xml.js
